Added a LOT of enemies
There are now A LOT more enemies to choose from to put within the game. Note that this is not all of them, simply the ones that caught my eye and would be useful to show by the end of the semester. Note that I went through and added all images, turned them into sprites, then made flipbooks from them. Some enemies lack things like an idle animation, but that is okay since I can just grab the sprites and make my own. Additionally, I left a folder titled "zzzzzz" which is simply an empty folder with 3 folders inside. This is 100% intentional and sped up the process for organizational purposes. Please do not delete this folder unless it is necessary and definitely let me know.